backend/requirements.txt
# Core dependencies
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
numpy==2.3.2
orjson==3.11.1
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
zstandard==0.23.0

# Database dependencies
sqlalchemy==2.0.43
alembic==1.16.4
psycopg2-binary==2.9.10

# Location services
geopy==2.4.1
googlemaps==4.10.0

# AI/ML dependencies  
langchain-core==0.3.74
langchain-openai==0.3.29
langchain-community==0.3.27
langsmith==0.4.13
openai==1.99.7

# Vector search and embeddings
sentence-transformers==2.2.2
faiss-cpu==1.8.0
chromadb==0.4.15

# Session management and caching
redis==3.5.3
redis-py-cluster==2.1.3

# Enhanced NLP
spacy==3.7.2
transformers==4.44.0

# Search and indexing
elasticsearch==8.10.1

# Real estate specific
beautifulsoup4==4.12.2
lxml==4.9.3

# Async support
asyncio-mqtt==0.13.0
aioredis==2.0.1

# Enhanced validation
marshmallow==3.20.1
jsonschema==4.19.1

# Background tasks
celery==5.3.1
kombu==5.3.1

# Monitoring and logging
prometheus-client==0.17.1
structlog==23.1.0

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
